refactor: moved pages/Anonymous.jsx to components directory since it's a component and not a page
